moved make_recipe tests to own file
